+++ config.go
@@ -1,7 +1,10 @@
 package sallust
 
 import (
+	"io/fs"
+	"net/url"
 	"os"
+	"path/filepath"
 
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
@@ -23,6 +26,12 @@ const (
 	// DefaultNameKey is the default value for EncoderConfig.NameKey.  This value
 	// is not used if EncoderConfig.DisableDefaultKeys is true.
 	DefaultNameKey = "name"
+
+	// Stdout is the reserved zap output path name that corresponds to stdout.
+	Stdout = "stdout"
+
+	// Stderr is the reserved zap output path name that corresponds to stderr.
+	Stderr = "stderr"
 )
 
 // EncoderConfig is an analog to zap.EncoderConfig.  This type is friendlier
@@ -250,7 +259,7 @@ type Config struct {
 	OutputPaths []string `json:"outputPaths" yaml:"outputPaths"`
 
 	// ErrorOutputPaths are the set of sinks for zap's internal messages.  This field
-	// corresponds to zap.Config.ErrorOutputPaths.  If unset, "stderr" is assumed.
+	// corresponds to zap.Config.ErrorOutputPaths.  If unset, Stderr is assumed.
 	//
 	// As with OutputPaths, environment variable references in each path are expanded
 	// unless DisablePathExpansion is true.
@@ -270,6 +279,11 @@ type Config struct {
 	// expansion, even with environment variables, is performed.
 	DisablePathExpansion bool `json:"disablePathExpansion" yaml:"disablePathExpansion"`
 
+	// Permissions is the optional nix-style file permissions to use when creating log files.
+	// If supplied, this value must be parseable via ParsePermissions.  If this field is unset,
+	// zap and lumberjack will control what permissions new log files have.
+	Permissions string `json:"permissions" yaml:"permissions"`
+
 	// Mapping is an optional strategy for expanding variables in output paths.
 	// If not supplied, os.Getenv is used.
 	Mapping func(string) string `json:"-" yaml:"-"`
@@ -286,11 +300,11 @@ func applyConfigDefaults(zc *zap.Config) {
 
 	if zc.Development && len(zc.OutputPaths) == 0 {
 		// NOTE: difference from zap ... in development they send output to stderr
-		zc.OutputPaths = []string{"stdout"}
+		zc.OutputPaths = []string{Stdout}
 	}
 
 	if len(zc.ErrorOutputPaths) == 0 {
-		zc.ErrorOutputPaths = []string{"stderr"}
+		zc.ErrorOutputPaths = []string{Stderr}
 	}
 
 	// NOTE: can't compare the Level with nil very easily, so just
@@ -299,9 +313,54 @@ func applyConfigDefaults(zc *zap.Config) {
 	zc.Level = zap.NewAtomicLevelAt(zapcore.InfoLevel)
 }
 
+// ensureExists makes sure the given path exists with the specified permissions.
+// If the path has already been created or if perms is 0, this function won't do anything.
+//
+// The path is treated as a URI in a similar fashion to zap.Open.
+func ensureExists(path string, perms fs.FileMode) (err error) {
+	if perms == 0 {
+		return
+	}
+
+	var f *os.File
+	defer func() {
+		if f != nil {
+			f.Close()
+		}
+	}()
+
+	switch {
+	case path == Stdout:
+		fallthrough
+
+	case path == Stderr:
+		break
+
+	// Windows hack:  filepath.Abs will return false outside of Windows
+	// for many paths.  This just makes sure we don't have to do a bunch
+	// of platform-specific nonsense.
+	case filepath.IsAbs(path):
+		f, err = os.OpenFile(path, os.O_CREATE|os.O_WRONLY, perms)
+
+	default:
+		var url *url.URL
+		url, err = url.Parse(path)
+		if err == nil {
+			f, err = os.OpenFile(url.Path, os.O_CREATE|os.O_WRONLY, perms)
+		}
+	}
+
+	return
+}
+
 // NewZapConfig creates a zap.Config enriched with features from these Options.
 // Primarily, this involves creating lumberjack URLs so that the registered sink
 // will create the appropriate infrastructure to do log file rotation.
+//
+// This method also enforces the Permissions field.  Any output or error path
+// will be created initially with the configured file permissions.  This allows
+// both zap's file sink and the custom lumberjack sink in this package to honor
+// custom permissions.
 func (c Config) NewZapConfig() (zc zap.Config, err error) {
 	zc = zap.Config{
 		Development:       c.Development,
@@ -334,6 +393,9 @@ func (c Config) NewZapConfig() (zc zap.Config, err error) {
 		}
 	}
 
+	var perms fs.FileMode
+	perms, err = ParsePermissions(c.Permissions)
+
 	if err == nil {
 		pt := PathTransformer{
 			Rotation: c.Rotation,
@@ -350,6 +412,16 @@ func (c Config) NewZapConfig() (zc zap.Config, err error) {
 		if err == nil {
 			zc.ErrorOutputPaths, err = ApplyTransform(pt.Transform, zc.ErrorOutputPaths...)
 		}
+	}
+
+	// Iterate over the transformed paths and ensure that any URIs that refer to
+	// files are created with relevant permissions.
+	for i := 0; err == nil && i < len(zc.OutputPaths); i++ {
+		err = ensureExists(zc.OutputPaths[i], perms)
+	}
+
+	for i := 0; err == nil && i < len(zc.ErrorOutputPaths); i++ {
+		err = ensureExists(zc.ErrorOutputPaths[i], perms)
 	}
 
 	if err == nil {